pyproject.toml
[project]
name = "yt-clipy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "textual>=1.0.0",
    "yt-dlp>=2024.12.23",
]
